requirements.txt
kafka-python==2.0.2
sendgrid==6.8.2
